POSSIBLE BREAKING GNSS CHANGE: LENA-R8 fix for GNSS operation, GNSS system types.
Most cellular modules with built-in GNSS chips accept a request to enable all GNSS system types in the AT+UGPS command when switching on GNSS and simply ignore the system types that cannot be supported, later returning a +UGIND URC to indicate which ones it has switched on.  SARA-R520M10 and LENA-R8xxxM10 do not: this was fixed for SARA-R520 when it was introduced in commit c881a877af63c787d5a99080e163ecc6f731c4b1, however the LENA-R8xxxM10 case is more complicated since LENA-R8xxxM10 does not support CMUX access to the built-in GNSS chip, only the more clunky AT+UGUBX access; the SARA-R520M10 fix did not work for that scenario.

With this commit a good set of system types (specifically not enabling BeiDou and Glonass at the same time) are cpnfigured for LENA-R8xxxM10.

The POSSIBLY BREAKING part of this change is that uGnssSetIntermediate() no longer returns an error if it is called when the GNSS transport type is U_GNSS_TRANSPORT_AT; in fact, in order to cope with the LENA-R8xxxM10 case, uGnssSetIntermediate() _should_ now be called when the GNSS transport type is U_GNSS_TRANSPORT_AT so that the GNSS code can determine the right sub-set of GNSS system types to request when bringing up a GNSS network on a cellular device.

Our thanks to mos216 for highlighting the issue.