Examples fix for cellular LENA-R8: include line endings in PEM certificates. (#1171)
The cellular LENA-R8 module requires line-endings in PEM-format files, the only cellular module to require this.  The certificate arrays used by the secure sockets examples did not include line endings and hence would fail.  This wasn't noticed because the original LENA-R8 FW version (likely the engineering sample version) used when the examples were first run did not have this requirement and, in the case of the examples, we only write the certificate files if the hash doesn't match, since that is the best usage pattern; this is not a hole in testing because the security credentials tests _do_ test writing the certificate files and the test certificates files used there _do_ include line endings in the PEM-format files, so all is good.

Our thanks to mos216 for highlighting this.