Linux change only: fix uPortQueueGetFree implementation. (#1169)
Calling fcntl(pQueue->fd[1], F_GETPIPE_SZ) will actually return the maximum size of the pipe for the given file descriptor, which is not what we are looking for here. Instead, it is safe to return the current number of free spaces based on the read count since it is protected.