POSSIBLY BREAKING HTTP Client change: chunked API. (#1177)
The HTTP Client API is extended, for cellular, to support chunked PUT, POST and GET operations, allowing the application to use smaller buffers for both uplink and downlink transfers.

The POSSIBLY BREAKING part of this is that the uHttpClientConnection_t structure, as passed to pUHttpClientOpen(), has an additional member, maxChunkLengthBytes, added on the end.  Provided you have assigned your connection structure to U_HTTP_CLIENT_CONNECTION_DEFAULT on creation, as recommended, all will be fine.