Cellular change for MQTT only: reset number of unread message on read fail. (#1176)
Some modules (e.g. LENA-R8) return an error if an attempt is made to read an MQTT message when there is none.  This is fine, but it might occur because the user has called uMqttClientGetUnread() and it, for some reason, has got out of step with reality (it is, after all, just tracking the unread message count inside the module, which cannot be queried directly) and so it makes sense to reset the number of unread messages to zero if a read fails.