Documentation change only: better advice on use of M10 in LENA-R8xxxM10.
The LENA-R8xxxM10 cellular module does not support access to the built-in M10 chip via CMUX, and access to the GNSS chip via ye olde AT+UGUBX mechanism, while possible, has limitations since there can be message collisions with the UPOS entity inside the cellular module [which cannot be stopped from talking to the GNSS chip if _it_ has switched it on] at the same time, leading to communications failures.  Hence it is advisable to use the separate, dedicated, TXD_GNSS/RXD_GNSS pins that the LENA-R8 module offers to access the internal GNSS chip and treat the M10 as a standalone GNSS device from a ubxlib perspective.

Advice is added to the location API and the examples concerning this.

Our apologies to mos216 for not adding this advice earlier.