BREAKING CHANGE FOR NRF5SDK PLATFORM: removal. (#1181)
The Nordic nRF5 SDK, deprecated some time ago by Nordic in favour of their new nRFConnect (i.e. Zephyr 3), the deprecation of which in ubxlib was announced at ubxlib release 1.3.0 back in August 2023, and which was deprecated in release 1.4.0 in March 2024, is now removed. If you wish to use ubxlib with a Nordic MCU please consider moving to nRFConnect.

Note on testing: the test instance that previously exercised nRF5SDK on an nRF52840 MCU (13.0) has been moved to test Zephyr instead but Zephyr is quite a lot bigger than nRF5SDK and that means that we cannot fit the test code for the cellular stuff into the available flash when using Zephyr, hence that test now only tests GNSS. This is not an issue, test coverage wise, as we have other instances that test Zephyr operation with cellular (e.g. instance 18, which runs Zephyr with SARA-R5 on an nRF5340 MCU).