Documentation update: IMPORTANT advice on uDeviceOpen().
Added some important advice on the pDeviceCfg parameter to uDeviceOpen(): you would normally define the uDeviceCfg_t structure pointed-to by pDeviceCfg as a const struct, so usually in flash, in which case the compiler will zero any non-populated members for you.  The additional advice is that, should you instead define what pDeviceCfg points to as a RAM variable BE SURE TO ZERO IT before populating it.  This will ensure compatibility with future extensions to the structure, and it is being mentioned now as it may be necessary to pass a run-time handle, established by the application, into the uDeviceOpen() function in the future, e.g. the new ESP-IDF I2C API might require a handle for an opened I2C bus to be passed down in the structure pointed-to by pDeviceCfg, requiring it to be in RAM.