GNSS change only: option for uDeviceClose() to call uGnssPwrOffBackup(). (#1187)
When uDeviceClose() is called with powerOff set to true on a GNSS device, the function uGnssPwrOff() is internally called to power the GNSS device down. This switches off the positioning engine but does not put the device into the lowest power state; the reason for this is that the mechanism through which the GNSS device is restored from back-up mode may require additional pins to be connected; e.g. toggling the I2C pins of the GNSS device will NOT cause it to return to normal operation (toggling the SPI or UART lines will).

With this commit the application can now set a Boolean flag, powerOffToBackup, in uDeviceCfgGnss_t; this will cause the device layer to call uGnssPwrOffBackup() instead of uGnssPwrOff(). It is up to the application to know how to restore the GNSS device to normal operation afterwards, e.g. it may be necessary to toggle RESET_N or connect a separate GPIO line; please refer to the section of the integration manual for your GNSS device that covers backup modes for details.