Cellular SARA-U201 change only: correct UART power saving support.
The behaviour of SARA-U201 when UART power saving is switched on was not handled properly: where it says in the AT manual "On SARA-U2 modules, when the AT+UPSV=1 command is issued by the DTE, the UART is immediately disabled" it really does mean _immediately_, there is not even an "OK" response.  Also, the module _immediately_ requires an "AT" or whatever to wake up, it doesn't wait the 6 seconds it has been asked to wait.

With this commit a new API is added to the AT client uAtClientWakeUpHandlerForce(), which tells the AT client that it has to do the "AT" thing to wake-up the module immediately, and the power-up sequence is modified so that if power saving is being switched on for a SARA-U201 it does not expect an "OK" response to the "AT+UPSV=1,1300" command.

Test: 32 cell