POSSIBLY BREAKING change to uPort layer: add GPIO interrupt support. (#1189)
* POSSIBLY BREAKING change to uPort layer: add GPIO interrupt support.

The GPIO portion of the port layer is updated to include GPIO interrupt support.  Specifically, three new fields are added to the uPortGpioConfig_t structure that is passed to uPortGpioConfig(), allowing an interrupt callback and either edge or level (where supported) triggering to be configured.  This is required for up-coming support of use of the Data Ready pin when reading from a GNSS device.

This change is backwards compatible with existing ports in that there is also a uPortGpioInterruptSupported() function with a WEAK implementation that returns false; the ubxlib code will not attempt to use GPIO interrupts on platforms where uPortGpioInterruptSupported() returns false.  Note that interrupts are not implemented for Windows and Linux (including Zephyr on Linux), since interrupts are not accessible on these platforms.

The POSSIBLY BREAKING part of the change is that, if anyone has used uPortGpioConfig() in their application and passed into it a uPortGpioConfig_t structure that has NOT been assigned to U_PORT_GPIO_CONFIG_DEFAULT, the three new fields will be left floating and anything may happen.  Please make sure that, if you have used uPortGpioConfig() in your application code, you have assigned your uPortGpioConfig_t to U_PORT_GPIO_CONFIG_DEFAULT on creation; all of the ubxlib code does this.