Internal naming fix in cellCfg: not just SARA any more. (#1193)
Some of the static functions in u_cell_cfg.c had "SaraR" in the name, because originally they applied only to SARA-R4/R5/R6.  However, things have moved on and the same functions now apply to LENA-R8, plus there are now LEXI-versions of the R422 and R52 modules, hence the naming has become actively confusing.

These static functions are renamed to remove "Sara".  The equivalent static functions for the SARA-U2 and LEXI-R10 modules remain named specifically, with SaraU2 and LexiR10 in their names.