GNSS change only: add support for Data Ready, AKA TXReady, pin. (#1201)
GNSS devices support toggling a TXReady pin when there is data available. This can be useful for power saving, allowing an MCU to sleep, or perform other tasks under the OS, while waiting for position from the GNSS device. Building on 128a72b4a2c63dfd67bbbe1c368ef8d80d415bb1, which added interrupt support to the GPIO API of the port layer, GNSS Data Ready is now supported using the fields pinDataReady and devicePioDataReady of the device configuration structure passed to uDeviceOpen() for a GNSS device. Note that pinDataReady was already present in the configuration structure, normally set to -1, but for this feature to work the PIO of the GNSS device that this MCU pin is connected to must also be known, hence the addition of devicePioDataReady, which must be set to a valid value if pinDataReady is not -1.

Note that this is only supported on M9 devices and later, on I2C and SPI interfaces, and also that devicePioDataReady must not already be in use for something else on the GNSS device: where that is the case uDeviceOpen() will return the error U_GNSS_ERROR_PIO_IN_USE.