Test change for instance 33 (LEXI-R10) only: APN setting.
The APN required for the "1NCE" SIM card we use with the LEXI-R10 test instance, instance 33, was missing from DATABASE.md, leading to instance 33 failing to register with the network after a module FW update, which likely erased a remembered APN setting from long ago.  The correct APN setting is now added for instance 33.